compatibility with napari 0.7.0
- Add model_config with arbitrary_types_allowed for pydantic v2 compat
  (SceneOverlay now inherits from pydantic v2 BaseModel, which cannot
  generate schemas for vispy ColorArray fields)
- Remove _add_overlay_to_visual calls and register overlay_to_visual
  mapping before adding to viewer._overlays (napari 0.7.0 auto-creates
  vispy overlays on dict insertion)
- Add **kwargs to VispyTimestampOverlay and VispyLayerAnnotatorOverlay
  __init__ (napari 0.7.0 passes font_manager/font_family kwargs)
- Remove overlay.events.text listener (text is a @property, not a
  field, so psygnal does not emit events for it)
- Replace .astype(int) with int() in render_as_rgb (dims.range values
  are now plain float, not numpy scalars)

+++ src/napari_timestamper/render_as_rgb.py
@@ -35,10 +35,10 @@ def render_as_rgb(
                 scale_factor=upsample_factor, flash=False
             ).shape
             rgb = np.zeros(
-                (viewer.dims.range[axis][1].astype(int) + 1, *target_shape),
+                (int(viewer.dims.range[axis][1]) + 1, *target_shape),
                 dtype=np.uint8,
             )
-            for j in range(viewer.dims.range[axis][1].astype(int) + 1):
+            for j in range(int(viewer.dims.range[axis][1]) + 1):
                 viewer.dims.set_current_step(axis, j)
                 rendered_img = viewer.export_figure(
                     scale_factor=upsample_factor, flash=False
@@ -51,13 +51,13 @@ def render_as_rgb(
             rgb = np.zeros(
                 (
                     len(axis),
-                    viewer.dims.range[axis[0]][1].astype(int) + 1,
+                    int(viewer.dims.range[axis[0]][1]) + 1,
                     *target_shape,
                 ),
                 dtype=np.uint8,
             )
             for ax in axis:
-                for j in range(viewer.dims.range[ax][1].astype(int) + 1):
+                for j in range(int(viewer.dims.range[ax][1]) + 1):
                     viewer.dims.set_current_step(ax, j)
                     rendered_img = viewer.export_figure(
                         scale_factor=upsample_factor, flash=False
